// assets/scripts/enums/GameEvent.ts
import { Enum } from "cc";

export enum GameEvent {
	NONE,

	REDIRECT_PROCESSING = 10,

	JOYSTICK_MOVE_START = 20,
	JOYSTICK_MOVE = 21,
	JOYSTICK_MOVE_END = 22,
	SET_INPUT_ENABLED = 23,

	CORRECT_VELOCITY = 30,

	CHECK_INTERACTION = 40,
	CHECK_INTERACTION_ONCE = 41,
	INTERACTION_START = 42,
	INTERACTION_END = 43,
	DISABLE_INTERACTIVE_AGENT = 44,

	REGISTER_BUTTON = 50,
	UNREGISTER_BUTTON = 51,

	CAMERA_TRANSITION = 60,
	CAMERA_SHAKE = 61,
	CAMERA_GET = 62,
	CAMERA_UI_GET = 63,
	CAMERA_UPDATE_POSITION = 64,
	CAMERA_FOCUS = 65,
	CAMERA_TRANSITION_TO_NODE = 66,
	CAMERA_SET_SETUP_IMMEDIATE = 67,
	CAMERA_INTRO_ROUTE = 68,

	STACK_FULL = 70,
	STACK_EMPTY = 71,
	STACK_ITEM_PLACED = 72,
	STACK_ITEM_REMOVED = 73,
	STACK_ADJUST_ON_FULL = 74,
	STACK_SORT = 75,
	STACK_REMOVE_TOP_ITEM = 76,
	STACK_ADD_NEW_SLOT = 77,
	STACK_PRE_FULL = 78,
	STACK_ITEM_BOX_SPAWN = 79,

	PUSH_TO_STACK = 80,
	POP_FROM_STACK = 81,
	TRANSMIT_TO_STACK = 82,
	FLY_FROM_STACK = 83,

	ENABLE_INPUT = 90,
	DISABLE_INPUT = 91,

	PLAYER_GET = 100,

	MOW_GRASS = 110,
	MOW_UPGRADE = 111,
	MOW_START = 112,
	MOW_END = 113,
	MOW = 114,
	MOW_CROP = 115,
	MOW_GRASS_TEST = 116,

	MOVER_START_MOVE = 120,
	MOVER_END_MOVE = 121,

	CUSTOMER_ARRIVE = 130,
	CUSTOMER_DEMAND = 131,
	CUSTOMER_DEMAND_SATISFY = 132,
	CUSTOMER_DEMAND_SATISFIED = 133,
	CUSTOMER_QUEUE_START = 134,
	CUSTOMER_LEAVE = 135,
	CUSTOMER_LEAVE_END = 136,
	CUSTOMER_LEAVE_START = 137,
	CUSTOMER_QUEUE = 138,
	CUSTOMER_DEMAND_SATISFIED_CAR = 139,

	PUNCH = 160,
	PUNCHABLE_NO_HEALTH = 161,
	PUNCHABLE_RESURRECT = 162,
	PUNCHING_START = 163,
	PUNCHING_END = 164,
	PUNCHING = 165,

	GET_TO_WORK = 170,

	APPEAR = 180,

	MONEY_ADD = 190,
	MONEY_SUB = 191,
	MONEY_GET = 192,
	PLAYER_MONEY_CHANGED = 193,

	GAME_END = 210,
	SHOW_END_SCREEN = 211,
	GAME_START = 212,
	ANALYTICS_SEND_EVENT = 213,
	ANALYTICS_SEND_COMPLETION_EVENT = 214,
	SCREEN_TOUCH_START = 215,

	TUTORIAL_SET_VALUE = 240,
	TUTORIAL_GET_VALUE = 241,
	TUTORIAL_ARROW_SET_VISIBILITY = 242,
	TUTORIAL_ENTER_HIDE_AREA = 243,
	TUTORIAL_LEAVE_HIDE_AREA = 244,
	TUTORIAL_ARROW_SET_TARGET = 245,

	SHOW_MAX_LABEL = 260,

	SOUND_PLAY_ONE_SHOT = 270,
	SOUND_SWITCH = 271,
	SOUND_BUTTON_PRESS = 272,
	SOUND_GET_IS_ON = 273,
	TOGGLE_SOUND = 274,

	HIDE_MAX_SIGN = 280,
	SHOW_MAX_SIGN = 281,

	GET_AGENT_ID = 290,

	COLLECT_RESOURCE = 300,

	ZONE_BUY = 310,
	ZONE_ENTER = 311,
	ZONE_EXIT = 312,

	PLAYER_MOVE = 330,
	PLAYER_SICKLE_TOGGLE_TRAIL = 331,
}

Enum(GameEvent);